requirements.txt
flask>=3.0
sqlalchemy>=2.0
apscheduler>=3.10
anthropic>=0.40
requests>=2.31
python-dotenv>=1.0
Pillow>=10.0
gunicorn>=22.0
